Update information in Legal, Home, Download, and Screenshots pages
This commit removes the references to JavaFX and Xenon in the Legal and Home Page and replaces image screenshots with a YouTube video in the Screenshots Page. It also removes some rows from the downloads table in the Downloads Page as they may be irrelevant or no longer in use. The improvements aim to enhance the clarity and relevance of the information provided on these pages.